Update to fix speed/state bug
the ESPHome firmware requires a state command to be sent along with a speed command.

However, this breaks the functionality for other fans that can't use that state command.

I made the ability to send the state command along with the speed command for those using the ESPHome firmware for their fans. The setting is optional and will default to false if not included in the config.